updated to spread out SignUp components for easier unit testing
