+++ src/pages/SignUp/SignUp.test.tsx
@@ -1,0 +1,14 @@
+import {
+    render,
+    screen,
+    fireEvent,
+    waitFor
+} from '@testing-library/react';
+import {SignUp} from '@/pages/SignUp/SignUp';
+
+describe.only('SignUp Component', () => {
+    test('should render', () => {
+	const {baseElement} = render(<SignUp />);
+        expect(baseElement).toBeDefined();
+    });
+});
