+++ src/pages/SignUp/index.tsx
@@ -1,0 +1,5 @@
+export * from './RoleSelect';
+export * from './AccountCredentials';
+export * from './LanguageModeSelect';
+export * from './LanguageInclusivitySelect';
+export * from './SignUp';
